functions/ingest-http/requirements.txt
google-cloud-storage==2.16.0
google-cloud-bigquery==3.17.2
functions-framework==3.5.0